+++ apps/backend/src/app.module.ts
@@ -3,10 +3,11 @@ import { AuthModule } from './auth/auth.module';
 import { HealthController } from './health/health.controller';
 import { MeController } from './me/me.controller';
 import { PrismaModule } from './prisma/prisma.module';
+import { QuizzesModule } from './quizzes/quizzes.module';
 import { UsersModule } from './users/users.module';
 
 @Module({
-  imports: [PrismaModule, AuthModule, UsersModule],
+  imports: [PrismaModule, AuthModule, UsersModule, QuizzesModule],
   controllers: [HealthController, MeController],
   providers: [],
 })